+++ packages/react/src/venn/layout/vennDiagramLayout.ts
@@ -7,70 +7,8 @@
 
 import { ISetCombinations, ISets } from '@upsetjs/model';
 import { circleIntersectionPoints } from './math';
+import { IArc, IArcSlice, ICircle, IUniverseSet } from './interfaces';
 
-export interface ICircle {
-  r: number;
-  x: number;
-  y: number;
-  align: 'center' | 'left' | 'right';
-}
-
-// could be slice
-export interface IArc {
-  rx: number;
-  ry: number;
-  rotation: number;
-  x2: number;
-  y2: number;
-  sweepFlag: boolean;
-  largeArcFlag: boolean;
-}
-
-export interface IArcSlice {
-  cx: number;
-  cy: number;
-
-  x1: number;
-  y1: number;
-  arcs: ReadonlyArray<IArc>;
-}
-
-export function generateArcSlicePath(slice: IArcSlice) {
-  return `M ${slice.x1},${slice.y1} ${slice.arcs
-    .map(
-      (arc) =>
-        `A ${arc.rx} ${arc.ry} ${arc.rotation} ${arc.largeArcFlag ? 1 : 0} ${arc.sweepFlag ? 1 : 0} ${arc.x2} ${arc.y2}`
-    )
-    .join(' ')}`;
-}
-
-export function generatePieSlice(c: ICircle, ratio: number, _secondary?: boolean) {
-  if (ratio <= 0) {
-    return '';
-  }
-  if (ratio >= 1) {
-    return `M ${c.x} ${c.y - c.r} A ${c.r} ${c.r} 0 1 0 ${c.x} ${c.y + c.r} A ${c.r} ${c.r} 0 1 0 ${c.x} ${c.y - c.r}`;
-  }
-  const pt = (angle: number) => ({
-    x: c.x + Math.cos(angle) * c.r,
-    y: c.y + Math.sin(angle) * c.r,
-  });
-  const deg2rad = Math.PI / 180;
-  const span = 360 * ratio;
-  const start = (c.align === 'center' ? 0 : c.align === 'left' ? 270 : 90) - 90 + (360 - span) / 2;
-  const end = start + span;
-
-  const startPt = pt(start * deg2rad);
-  const endPt = pt(end * deg2rad);
-  return `M ${c.x} ${c.y} L ${startPt.x} ${startPt.y} A ${c.r} ${c.r} 0 ${ratio > 0.5 ? 1 : 0} 1 ${endPt.x} ${
-    endPt.y
-  } Z`;
-}
-
-export interface IUniverseSet extends IArcSlice {
-  width: number;
-  height: number;
-}
 // could be slice of three
 
 export function generateUniverseSetPath(l: IUniverseSet) {
